Divide the fuzzers into packages with logical grouping
