+++ test/fuzz/client-go/fuzz.go
@@ -1,0 +1,38 @@
+//go:build gofuzz
+// +build gofuzz
+
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package clientgo
+
+import (
+	"k8s.io/client-go/util/keyutil"
+)
+
+// FuzzParsePrivateKeyPEM implements a fuzzer
+// that targets keyutil.ParsePrivateKeyPEM
+func FuzzParsePrivateKeyPEM(data []byte) int {
+	_, _ = keyutil.ParsePrivateKeyPEM(data)
+	return 1
+}
+
+// FuzzParsePublicKeysPEM implements a fuzzer
+// that targets keyutil.ParsePublicKeysPEM
+func FuzzParsePublicKeysPEM(data []byte) int {
+	_, _ = keyutil.ParsePublicKeysPEM(data)
+	return 1
+}
