+++ test/fuzz/cmd/fuzz.go
@@ -1,0 +1,31 @@
+//go:build gofuzz
+// +build gofuzz
+
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cmd
+
+import (
+	kubeadmutil "k8s.io/kubernetes/cmd/kubeadm/app/util"
+)
+
+// FuzzParseHostPort implements a fuzzer
+// that targets kubeadmutil.ParseHostPort
+func FuzzParseHostPort(data []byte) int {
+	_, _, _ = kubeadmutil.ParseHostPort(string(data))
+	return 1
+}
